Fixed bug introduced by membership validation optimization
The bug caused verification of deleted memberships to erroneously report as failed, because the check was still assuming we were looking at live group data, when we were actually looking at unmodified local user data. Changed this to grab a new copy of the live user data to check against that.